Clean up template tree modifier plugins
